+++ OmneFictio.WebApi/Dtos/RateDto.cs
@@ -2,6 +2,6 @@
 namespace OmneFictio.WebApi.Dtos;
 public partial class RateDto
 {
-    public double body { get; set; }
+    public byte body { get; set; }
 }
