+++ OmneFictio.WebApi/Entities/Rate.cs
@@ -8,7 +8,7 @@ namespace OmneFictio.WebApi.Entities
         public int id { get; set; }
         public int accountId { get; set; }
         public int postId { get; set; }
-        public double body { get; set; }
+        public byte body { get; set; }
 
         public virtual Account account { get; set; } = null!;
         public virtual Post post { get; set; } = null!;
